+++ support/cas-server-support-cassandra-ticket-registry/src/main/java/org/apereo/cas/ticket/registry/NoSqlTicketRegistryDao.java
@@ -1,4 +1,4 @@
-package org.apereo.cas.dao;
+package org.apereo.cas.ticket.registry;
 
 import org.apereo.cas.ticket.Ticket;
 import org.apereo.cas.ticket.TicketGrantingTicket;
